+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Improvements
 * Core: Added an SEO-item in the advanced-settings-section. For now only noodp and noydir are implemented.
 * Core: Added advanced SEO-settings in the SEO-tab (index,follow).
 * Core: Added a setting to use no-follow on links inside user-comments.
+* Core: If Google Analytics is available, all outgoing links will be tracked by eventtracking.
 
 
 2.5.2 (2011-07-19)